+++ src/main/java/com/jinddung2/givemeticon/domain/brand/controller/dto/request/BrandUpdateNameRequest.java
@@ -1,0 +1,6 @@
+package com.jinddung2.givemeticon.domain.brand.controller.dto.request;
+
+public record BrandUpdateNameRequest(
+        String name
+) {
+}
