--- src/main/java/com/jinddung2/givemeticon/domain/brand/dto/request/BrandUpdateNameRequest.java
@@ -1,6 +1,0 @@
-package com.jinddung2.givemeticon.domain.brand.dto.request;
-
-public record BrandUpdateNameRequest(
-        String name
-) {
-}
