+++ src/main/java/com/jinddung2/givemeticon/domain/category/controller/request/CategoryUpdateNameRequest.java
@@ -1,0 +1,6 @@
+package com.jinddung2.givemeticon.domain.category.controller.request;
+
+public record CategoryUpdateNameRequest(
+        String name
+) {
+}
