--- src/main/java/com/jinddung2/givemeticon/domain/category/dto/request/CategoryUpdateNameRequest.java
@@ -1,6 +1,0 @@
-package com.jinddung2.givemeticon.domain.category.dto.request;
-
-public record CategoryUpdateNameRequest(
-        String name
-) {
-}
