+++ src/main/java/com/jinddung2/givemeticon/domain/item/controller/dto/ItemDto.java
@@ -1,4 +1,4 @@
-package com.jinddung2.givemeticon.domain.item.dto;
+package com.jinddung2.givemeticon.domain.item.controller.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.jinddung2.givemeticon.domain.item.domain.Item;
