+++ src/main/java/com/jinddung2/givemeticon/domain/user/controller/dto/request/PasswordResetRequest.java
@@ -1,0 +1,6 @@
+package com.jinddung2.givemeticon.domain.user.controller.dto.request;
+
+public record PasswordResetRequest(
+        String email
+) {
+}
