--- src/main/java/com/jinddung2/givemeticon/domain/user/dto/request/PasswordResetRequest.java
@@ -1,6 +1,0 @@
-package com.jinddung2.givemeticon.domain.user.dto.request;
-
-public record PasswordResetRequest(
-        String email
-) {
-}
