+++ src/test/java/com/jinddung2/givemeticon/domain/category/controller/CategoryControllerTest.java
@@ -3,8 +3,8 @@ package com.jinddung2.givemeticon.domain.category.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.jinddung2.givemeticon.common.config.WebConfig;
 import com.jinddung2.givemeticon.common.security.interceptor.AuthInterceptor;
+import com.jinddung2.givemeticon.domain.category.controller.request.CategoryUpdateNameRequest;
 import com.jinddung2.givemeticon.domain.category.domain.Category;
-import com.jinddung2.givemeticon.domain.category.dto.request.CategoryUpdateNameRequest;
 import com.jinddung2.givemeticon.domain.category.exception.NotFoundCategoryException;
 import com.jinddung2.givemeticon.domain.category.service.CategoryService;
 import com.jinddung2.givemeticon.domain.user.service.LoginService;
