+++ src/test/java/com/jinddung2/givemeticon/domain/item/facade/ItemCreationFacadeTest.java
@@ -1,10 +1,10 @@
 package com.jinddung2.givemeticon.domain.item.facade;
 
+import com.jinddung2.givemeticon.domain.brand.controller.dto.BrandDto;
 import com.jinddung2.givemeticon.domain.brand.domain.Brand;
-import com.jinddung2.givemeticon.domain.brand.dto.BrandDto;
 import com.jinddung2.givemeticon.domain.brand.service.BrandService;
+import com.jinddung2.givemeticon.domain.item.controller.dto.request.ItemCreateRequest;
 import com.jinddung2.givemeticon.domain.item.domain.Item;
-import com.jinddung2.givemeticon.domain.item.dto.request.ItemCreateRequest;
 import com.jinddung2.givemeticon.domain.item.service.ItemService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
