+++ src/test/java/com/jinddung2/givemeticon/domain/user/controller/facade/CreateAccountFacadeTest.java
@@ -1,9 +1,9 @@
 package com.jinddung2.givemeticon.domain.user.controller.facade;
 
-import com.jinddung2.givemeticon.domain.account.dto.request.CreateAccountRequest;
+import com.jinddung2.givemeticon.domain.account.request.CreateAccountRequest;
 import com.jinddung2.givemeticon.domain.account.service.AccountService;
 import com.jinddung2.givemeticon.domain.user.domain.User;
-import com.jinddung2.givemeticon.domain.user.presentation.facade.CreateAccountFacade;
+import com.jinddung2.givemeticon.domain.user.facade.CreateAccountFacade;
 import com.jinddung2.givemeticon.domain.user.service.UserService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
